requirements.txt
streamlit==1.20.0
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
opencv-python==4.7.0.72
scikit-image==0.20.0
scikit-learn==1.2.2
joblib==1.2.0
Pillow==9.5.0
mahotas==1.4.13